Adds support for deserializing response and filtering
