+++ Cargo.toml
@@ -6,5 +6,6 @@ edition = "2024"
 [dependencies]
 clap = { version = "4.5.32", features = ["derive"] }
 derive_more = { version = "2.0.1", features = ["from"] }
-reqwest = "0.12.12"
+reqwest = { version = "0.12.12", features = ["json"] }
+serde = { version = "1.0.219", features = ["derive"] }
 tokio = { version = "1.44.0", features = ["full"] }
